+++ imports/scenes/Contract/ClientSetup/index.js
@@ -5,14 +5,20 @@ import tools from '/imports/startup/tools/index';
 import Box from '/imports/components/Box/index';
 import SuggestionBar from '/imports/components/SuggestionBar/index';
 import RegisterData from '/imports/components/RegisterData/index';
+import ConfirmationWindow from '/imports/components/ConfirmationWindow/index';
+import FooterButtons from '/imports/components/FooterButtons/index';
 
 export default class ClientSetup extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
       clientId: '',
-      clientWindow: false
+      clientWindow: false,
+      confirmationWindow: false
     }
+  }
+  toggleConfirmationWindow = () => {
+    this.setState({ confirmationWindow: !this.state.confirmationWindow })
   }
   handleChange = (e) => {
     this.setState({ clientId: e.target.value })
@@ -21,6 +27,18 @@ export default class ClientSetup extends React.Component {
     var clientWindow = this.state.clientWindow ? false : {};
     this.setState({ clientWindow })
   }
+  verification = () => {
+    if (!this.state.clientId) {
+      this.toggleConfirmationWindow();
+    } else {
+      this.saveEdits();
+    }
+  }
+  saveEdits = () => {
+    this.props.updateContract({
+      clientId: this.state.clientId
+    }, this.props.closeWindow)
+  }
   render () {
     if (!this.props.proposal) return null;
     return (
@@ -28,7 +46,7 @@ export default class ClientSetup extends React.Component {
         title="Seleção Inicial de Cliente">
           <div className="client-setup">
             <h4 className="client-setup__proposal-title">
-              Dados da Proposta
+              Dados da Proposta:
             </h4>
             <div className="client-setup__proposal-body">
               <div>
@@ -48,7 +66,7 @@ export default class ClientSetup extends React.Component {
                 {this.props.proposal.client.phone}
               </div>
             </div>
-            <div className="client-setup__existing-client-body">
+            <div className="client-setup__client-body">
               <SuggestionBar
                 title="Selecionar Cliente:"
                 name="clientId"
@@ -58,16 +76,25 @@ export default class ClientSetup extends React.Component {
                 onClick={this.handleChange}>
               </SuggestionBar>
               <button onClick={this.toggleClientWindow} className="button--pill client-setup__add-new-button">
-                ADICIONAR NOVO CLIENTE
+                +
               </button>
-              {this.state.clientWindow ?
-                <RegisterData
-                  type="clients"
-                  item={this.state.clientWindow}
-                  toggleWindow={this.toggleClientWindow}
-                />
-              : null}
             </div>
+            <FooterButtons buttons={[
+              {text: "Adicionar Cliente", onClick: this.verification}
+            ]}/>
+            {this.state.clientWindow ?
+              <RegisterData
+                type="clients"
+                item={this.state.clientWindow}
+                toggleWindow={this.toggleClientWindow}
+              />
+            : null}
+            <ConfirmationWindow
+              isOpen={this.state.confirmationWindow}
+              closeBox={this.toggleConfirmationWindow}
+              message="Deseja mesmo iniciar sem cadastrar o cliente?"
+              leftButton={{text: "Não", className: "button--secondary", onClick: this.toggleConfirmationWindow}}
+              rightButton={{text: "Sim", className: "button--danger", onClick: this.saveEdits}}/>
           </div>
       </Box>
     )